renamed coregistration_utilities module
