--- docs/source/eolearn.coregistration.coregistration_utilities.rst
@@ -1,7 +1,0 @@
-eolearn.coregistration.coregistration\_utilities
-================================================
-
-.. automodule:: eolearn.coregistration.coregistration_utilities
-    :members:
-    :undoc-members:
-    :show-inheritance:
