+++ docs/source/eolearn.coregistration.utils.rst
@@ -1,0 +1,7 @@
+eolearn.coregistration.utils
+============================
+
+.. automodule:: eolearn.coregistration.utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
